+++ packages/block-editor/src/components/inserter/test/block-types-tab.native.js
@@ -49,7 +49,8 @@ describe( 'BlockTypesTab component', () => {
 		selectMock.getInserterItems.mockReturnValue( items );
 
 		const blockItems = items.filter(
-			( { category } ) => category !== 'reusable'
+			( { id, category } ) =>
+				category !== 'reusable' && id !== 'core-embed/a-paragraph-embed'
 		);
 		const component = shallow(
 			<BlockTypesTab